Moved file with data
